♻️(api) move health_huey view to presentation.api,
