+++ src/tycho/presentation/api/urls.py
@@ -5,8 +5,13 @@ from rest_framework_simplejwt.views import (
     TokenRefreshView,
 )
 
+from presentation.api.views import HueyHealthView
+
+app_name = "api"
+
 urlpatterns = [
     path("token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
     path("token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
+    path("health/huey/", HueyHealthView.as_view(), name="health_huey"),
     path("schema/", SpectacularAPIView.as_view(), name="schema"),
 ]
